Assemble uint32_t instead of uint8_t
Vulkan receives SPIR-V modules with a uint32_t alignment. Returning
uint8_t forced users to invoke undefined behaviour (reinterpret_cast)
or copy.

+++ include/sirit/sirit.h
@@ -37,7 +37,7 @@ public:
      * externally.
      * @return A stream of bytes representing a SPIR-V module.
      */
-    std::vector<std::uint8_t> Assemble() const;
+    std::vector<std::uint32_t> Assemble() const;
 
     /// Adds a SPIR-V extension.
     void AddExtension(std::string extension_name);
